Fix bug in available memory computation
Prior to this commit, there was a potential bug:

The DeviceCachingAllocator gets the device's allocation limit using
cudaMemGetInfo. This function gets the memory for the current context.

AFAICT there is nothing to guarantee that the DeviceCachingAllocator
being used corresponds to the currently active context. So it is
possible that the memory limit for the device will be incorrect.

cudaGetDeviceProperties function appears to be a better choice:
1. It allows specifying the device that should be queried
2. It does not initialize/switch the cuda context
3. In my local testing, I found it to be faster than cudaMemGetInfo

This commit also eliminates the need for CUDACachingAllocator to
call SetDevice(...) prior to getting/setting the memory fraction.